requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
supabase==2.0.3
boto3==1.29.3
selenium==4.15.2
webdriver-manager==4.0.1 
asyncpg==0.28.0
botocore==1.32.3
jmespath==1.0.1
s3transfer==0.7.0
requests==2.31.0
aiohttp==3.8.6
schedule==1.2.1